Fix dead links on "tabs" and "list" pages (#1301)
* 🐛  Fix dead link by moving example

* 🚨  Reorder imports

* 🚨 Reorder imports

* 🐛  Update imports

* 🐛  Fix list dead link by moving to own folder

* 🐛  Place list-swipe ex in folder to fix dead link

* 🐛  Fix reorder-list and no-shape dead links

* 🐛  Fix dead tabs link

* 🚨  reorder imports

* ♻️ Move list-example module to list-example folder

* 🤦‍♂️ i forgot to change gitUrl back to master

* ♻️  Remove BaseListComponent duplicates

* ♻️ Rename .../list-example/components to examples
