+++ apps/cookbook/src/app/examples/examples.routes.ts
@@ -22,18 +22,18 @@ import { FormFieldExampleComponent } from './form-field-example/form-field-examp
 import { GridExampleComponent } from './grid-example/grid-example.component';
 import { IconExampleComponent } from './icon-example/icon-example.component';
 import { ItemExampleComponent } from './item-example/item-example.component';
-import { ListBoldTextOnRowSelectionExampleComponent } from './list/components/bold-text-on-row-selection';
-import { ListColoredItemsExampleComponent } from './list/components/colored-items';
-import { ListWithDividersExampleComponent } from './list/components/dividers';
-import { ListWithHeaderAndFooterExampleComponent } from './list/components/header-and-footer';
-import { ListItemsExampleComponent } from './list/components/items';
-import { ListLoadOnDemandExampleComponent } from './list/components/load-on-demand/list-load-on-demand-example.component';
-import { ListNoShapeExampleComponent } from './list/components/no-shape/list-no-shape-example.component';
-import { ListWithSectionsExampleComponent } from './list/components/sections';
-import { ListWithSectionsAndColoredItemsExampleComponent } from './list/components/sections-and-colored-items';
-import { ListSelectableItemsExampleComponent } from './list/components/selectable-items';
-import { ListSwipeExampleComponent } from './list/components/swipe/list-swipe-example.component';
-import { ListExampleComponent } from './list/list-example.component';
+import { ListBoldTextOnRowSelectionExampleComponent } from './list-example/examples/bold-text-on-row-selection';
+import { ListColoredItemsExampleComponent } from './list-example/examples/colored-items';
+import { ListWithDividersExampleComponent } from './list-example/examples/dividers';
+import { ListWithHeaderAndFooterExampleComponent } from './list-example/examples/header-and-footer';
+import { ListItemsExampleComponent } from './list-example/examples/items';
+import { ListWithSectionsExampleComponent } from './list-example/examples/sections';
+import { ListWithSectionsAndColoredItemsExampleComponent } from './list-example/examples/sections-and-colored-items';
+import { ListSelectableItemsExampleComponent } from './list-example/examples/selectable-items';
+import { ListExampleComponent } from './list-example/list-example.component';
+import { ListLoadOnDemandExampleComponent } from './list-load-on-demand-example/list-load-on-demand-example.component';
+import { ListNoShapeExampleComponent } from './list-no-shape-example/list-no-shape-example.component';
+import { ListSwipeExampleComponent } from './list-swipe-example/list-swipe-example.component';
 import { LoadingOverlayExampleComponent } from './loading-overlay-example/loading-overlay-example.component';
 import { ModalExampleComponent } from './modal-example/modal-example.component';
 import { ModalRoutePage1ExampleComponent } from './modal-example/modal-route-example/modal-route-page1-example.component';
@@ -48,14 +48,14 @@ import { PageFixedTitleAndActionsExampleComponent } from './page-example/fixed-t
 import { PageSimpleExampleComponent } from './page-example/simple/page-simple-example.component';
 import { ProgressCircleExampleComponent } from './progress-circle-example/progress-circle-example.component';
 import { RadioExampleComponent } from './radio-example/radio-example.component';
-import { ReorderListExampleComponent } from './reorder-list/reorder-list-example.component';
+import { ReorderListExampleComponent } from './reorder-list-example/reorder-list-example.component';
 import { SegmentedControlExampleComponent } from './segmented-control-example/segmented-control-example.component';
 import { SlideButtonExampleComponent } from './slide-button-example/slide-button-example.component';
 import { SlidesExampleComponent } from './slides-example/slides-example.component';
 import { SpinnerExampleComponent } from './spinner-example/spinner-example.component';
 import { StockChartExampleComponent } from './stock-chart-example/stock-chart-example.component';
-import { TabExampleComponent } from './tabs/tab/tab-example.component';
-import { TabsExampleComponent } from './tabs/tabs-example.component';
+import { TabExampleComponent } from './tabs-example/tab/tab-example.component';
+import { TabsExampleComponent } from './tabs-example/tabs-example.component';
 import { ToastExampleComponent } from './toast-example/toast-example.component';
 import { ToggleButtonExampleComponent } from './toggle-button-example/toggle-button-example.component';
 import { ToggleExampleComponent } from './toggle-example/toggle-example.component';
