+++ apps/cookbook/src/app/examples/list-example/examples/bold-text-on-row-selection.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from './base-list.component';
+import { BaseListComponent } from '../../list-shared/base-list.component';
 
 export const ListBoldTextOnRowSelectionExampleTemplate = `<kirby-list [items]="items" (itemSelect)="onItemSelect($event)">
   <kirby-item *kirbyListItemTemplate="let item">
