+++ apps/cookbook/src/app/examples/list-example/examples/dividers.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from './base-list.component';
+import { BaseListComponent } from '../../list-shared/base-list.component';
 
 export const ListWithDividersExampleTemplate = `<kirby-list [items]="items" showDivider="true">
   <kirby-item *kirbyListItemTemplate="let item">
