+++ apps/cookbook/src/app/examples/list-example/examples/sections-and-colored-items.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { BaseListComponent } from './base-list.component';
+import { BaseListComponent } from '../../list-shared/base-list.component';
 
 export const ListWithSectionsAndColoredItemsExampleTemplate = `<kirby-list
   [items]="items"
