+++ apps/cookbook/src/app/examples/list-example/list-example.component.scss
