+++ apps/cookbook/src/app/examples/list-example/list-example.module.ts
@@ -1,0 +1,31 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+import { KirbyModule } from '@kirbydesign/designsystem';
+
+import { ListBoldTextOnRowSelectionExampleComponent } from './examples/bold-text-on-row-selection';
+import { ListColoredItemsExampleComponent } from './examples/colored-items';
+import { ListWithDividersExampleComponent } from './examples/dividers';
+import { ListWithHeaderAndFooterExampleComponent } from './examples/header-and-footer';
+import { ListItemsExampleComponent } from './examples/items';
+import { ListWithSectionsExampleComponent } from './examples/sections';
+import { ListWithSectionsAndColoredItemsExampleComponent } from './examples/sections-and-colored-items';
+import { ListSelectableItemsExampleComponent } from './examples/selectable-items';
+
+const listExamples = [
+  ListBoldTextOnRowSelectionExampleComponent,
+  ListWithDividersExampleComponent,
+  ListSelectableItemsExampleComponent,
+  ListColoredItemsExampleComponent,
+  ListWithSectionsAndColoredItemsExampleComponent,
+  ListWithHeaderAndFooterExampleComponent,
+  ListWithSectionsExampleComponent,
+  ListItemsExampleComponent,
+];
+
+@NgModule({
+  imports: [CommonModule, KirbyModule],
+  declarations: [...listExamples],
+  exports: [...listExamples],
+})
+export class ListExamplesModule {}
