+++ apps/cookbook/src/app/examples/list-no-shape-example/list-no-shape-example.component.scss
