+++ apps/cookbook/src/app/examples/list-shared/base-list.component.ts
