+++ apps/cookbook/src/app/examples/list-swipe-example/list-swipe-example.component.scss
