+++ apps/cookbook/src/app/examples/tabs-example/tab/tab-example.component.ts
