+++ apps/cookbook/src/app/examples/tabs-example/tabs-example.component.scss
