+++ apps/cookbook/src/app/showcase/list-showcase/list-showcase.component.ts
@@ -1,13 +1,13 @@
 import { Component, OnInit } from '@angular/core';
+import { ListColoredItemsExampleTemplate } from '~/app/examples/list-example/examples/colored-items';
+import { ListWithDividersExampleTemplate } from '~/app/examples/list-example/examples/dividers';
+import { ListWithHeaderAndFooterExampleTemplate } from '~/app/examples/list-example/examples/header-and-footer';
+import { ListWithSectionsExampleTemplate } from '~/app/examples/list-example/examples/sections';
+import { ListWithSectionsAndColoredItemsExampleTemplate } from '~/app/examples/list-example/examples/sections-and-colored-items';
+import { ListSelectableItemsExampleTemplate } from '~/app/examples/list-example/examples/selectable-items';
 
-import { ListBoldTextOnRowSelectionExampleTemplate } from '../../examples/list/components/bold-text-on-row-selection';
-import { ListWithDividersExampleTemplate } from '~/app/examples/list/components/dividers';
-import { ListSelectableItemsExampleTemplate } from '~/app/examples/list/components/selectable-items';
-import { ListColoredItemsExampleTemplate } from '~/app/examples/list/components/colored-items';
-import { ListWithSectionsAndColoredItemsExampleTemplate } from '~/app/examples/list/components/sections-and-colored-items';
-import { ListWithHeaderAndFooterExampleTemplate } from '~/app/examples/list/components/header-and-footer';
-import { ListWithSectionsExampleTemplate } from '~/app/examples/list/components/sections';
-import { ListItemsExampleTemplate } from '../../examples/list/components/items';
+import { ListBoldTextOnRowSelectionExampleTemplate } from '../../examples/list-example/examples/bold-text-on-row-selection';
+import { ListItemsExampleTemplate } from '../../examples/list-example/examples/items';
 
 declare var require: any;
 
@@ -26,7 +26,7 @@ export class ListShowcaseComponent implements OnInit {
   sectionsExampleTemplate: string = ListWithSectionsExampleTemplate;
   items: string = ListItemsExampleTemplate;
 
-  exampleHtml: string = require('!raw-loader!../../examples/list/list-example.component.html')
+  exampleHtml: string = require('!raw-loader!../../examples/list-example/list-example.component.html')
     .default;
 
   sectionHeaderExampleHtml: string =
