+++ apps/cookbook/src/app/showcase/list-swipe-showcase/list-swipe-showcase.component.ts
@@ -7,6 +7,6 @@ declare var require: any;
   styleUrls: ['./list-swipe-showcase.component.scss'],
 })
 export class ListSwipeShowcaseComponent {
-  exampleHtml: string = require('!raw-loader!../../examples/list/components/swipe/list-swipe-example.component.html')
+  exampleHtml: string = require('!raw-loader!../../examples/list-swipe-example/list-swipe-example.component.html')
     .default;
 }
